Move URL tests to a separate file; a few test for equality / inequality
